pyproject.toml
[project]
name = "trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.68.0",
    "cohere>=5.18.0",
    "glob2>=0.7",
    "llama-index>=0.14.2",
    "llama-index-embeddings-cohere>=0.6.1",
    "llama-index-llms-cohere>=0.6.1",
    "llama-index-storage-docstore-mongodb>=0.4.1",
    "llama-index-storage-kvstore-mongodb>=0.4.1",
    "llama-index-vector-stores-mongodb>=0.8.1",
    "llama-index-vector-stores-weaviate>=1.4.1",
    "notebook>=7.4.5",
    "pdfplumber>=0.11.7",
    "pymongo>=4.15.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "weaviate-client>=4.16.10",
]
